rfct: create utils and extract filter_contours_area_of_image_tables
